KStream/KTable implementation of basic "Count" function
Challenges:
- Both the derived KStream + KTable have batching. Data isn't calculated second by second.

Reflections:
- If derived streams were a non-negotiable this would have implication for FE design + websocket service design
- Might need to reconsider use of Kstream/KTable for this project

+++ server/src/main/java/com/jdsg/datalogger/dto/ChartData.java
@@ -6,7 +6,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.time.LocalTime;
 
 @JsonDeserialize
 @JsonSerialize
@@ -14,7 +13,7 @@ import java.time.LocalTime;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ChartData {
-    private int value;
+    private Double value;
     private String sensorId;
     private String timeStamp;
 }
